+++ src/main/resources/static/style/style.css
@@ -249,7 +249,7 @@ input.btn.btn-primary {
 
 /* Login CSS End ---------------------------------------------------------*/
 
-/* *************************************ABOUT US************************************** */
+/* *************************************ORIGINAL ABOUTUS CSS************************************** */
 
 .flex {
   display: flex;
@@ -288,4 +288,25 @@ padding:10px;
 .pic {
   max-width:25%;
   max-height: 25%;
+}
+
+*************************
+social-link {
+  width: 30px;
+  height: 30px;
+  border: 1px solid #ddd;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  color: #666;
+  border-radius: 50%;
+  transition: all 0.3s;
+  font-size: 0.9rem;
+}
+
+.social-link:hover,
+.social-link:focus {
+  background: #ddd;
+  text-decoration: none;
+  color: #555;
 }
