+++ src/main/resources/templates/aboutus.html
@@ -8,68 +8,67 @@
 
 <body>
 <nav th:replace="fragments/common :: navbar"></nav>
-<div class="frame" class=flex>
+<div class=flex>
     <a id="home-button-aboutus" class="nav-button" href="/" target="_self">Home</a>
     <br><br><br>
 
     <div class="dev-content">
         <h1>Meet the Team</h1>
         <br>
-        <p class=readable>We have created a ordering website!!
-        </p>
+        <p class=readable>We have created a ordering website!!</p>
         <br>
     </div>
 
     <br>
 
     <div class=profile-img>
-        <img src="/styles/pics/rachael.jpg" class="pic">
+        <img src="./style/pics/ellen.jpg" class="pic">
     </div>
     <div>
 
         <p class=bold-name>Ellen Conley <a href="https://github.com/egconley" target="_blank"><img
-                src="/styles/pics/github_PNG58.png" class="github-logo"></a><a
-                href="https://www.linkedin.com/in/ellenconley/" target="_blank"><img src="/styles/pics/li.png"
-                                                                                    class="github-logo"></a></p>
+                src="./style/pics/github_PNG58.png" class="github-logo"></a><a
+                href="https://www.linkedin.com/in/ellenconley/" target="_blank"><img src="./style/pics/li.png"
+                                                                                     class="github-logo"></a></p>
         <p class=readable>
-            Hi! I am.........
+            Full Stack Javascript and Java Developer with a background in data analysis. Analytical thinker with a passion for problem solving and an innate sense of curiosity, openness, and adaptability.
         </p><br>
     </div>
 
     <div class=profile-img>
-        <img src="/styles/pics/hai.jpg" class="pic">
+        <img src="./style/pics/hai.png" class="pic">
     </div>
     <div>
         <p class=bold-name>Hai Le<a href="https://github.com/haitle16" target="_blank"><img
-                src="/styles/pics/github_PNG58.png" class="github-logo"></a><a
-                href="https://www.linkedin.com/in/haitle16/" target="_blank"><img src="/styles/pics/li.png"
-                                                                                                   class="github-logo"></a></p>
+                src="./style/pics/github_PNG58.png" class="github-logo"></a><a
+                href="https://www.linkedin.com/in/haitle16/" target="_blank"><img src="./style/pics/li.png"
+                                                                                  class="github-logo"></a></p>
         <p class=readable>
             My name is .......
         </p><br>
     </div>
 
     <div class=profile-img>
-        <img src="/styles/pics/rachael.jpeg" class="pic">
+        <img src="./style/pics/rachael.jpeg" class="pic">
     </div>
     <div>
 
         <p class=bold-name>Rachael Rice <a href="https://github.com/Rachnicrice" target="_blank"><img
-                src="/styles/pics/github_PNG58.png" class="github-logo"></a><a href="https://www.linkedin.com/in/rachaelrice/"
-                                                                               target="_blank"><img src="/styles/pics/li.png" class="github-logo"></a></p>
+                src="./style/pics/github_PNG58.png" class="github-logo"></a><a href="https://www.linkedin.com/in/rachaelrice/"
+                                                                               target="_blank"><img src="./style/pics/li.png" class="github-logo"></a></p>
         <p class=readable>
             Rachael, a once customer service expert turned software developer, was inspired to create Service Express through her own experience as a sales rep. She is dedicated to creating solutions through client-focused design and by prioritizing user experience. When she is not studying data structures and algorithms, you can find her binge watching Netflix shows with her partner Sean.
         </p>
     </div>
 
     <div class=profile-img>
-        <img src="/styles/pics/shingo.jpeg" class="pic">
+        <img src="./style/pics/shingo.JPG" class="pic">
     </div>
     <div>
 
         <p class=bold-name>Shingo Nakajima <a href="https://github.com/shingoed" target="_blank"><img
-                src="/styles/pics/github_PNG58.png" class="github-logo"></a><a href="https://www.linkedin.com/in/shingoed/"
-                                                                               target="_blank"><img src="/styles/pics/li.png" class="github-logo"></a></p>
+                src="./style/pics/github_PNG58.png" class="github-logo"></a><a href="https://www.linkedin.com/in/shingoed/"
+                                                                               target="_blank"><img src="./style/pics/li.png" class="github-logo"></a></p>
         <p class=readable>
             Shingo, a passionate coder coming from Code Fellows bootcamp with expertise in agile development methodology and experience with software development using continuous integration/deployment pipeline. A motivated team player who is highly proficient in Java and well versed in JavaScript with experience building Android apps.A passionate coder coming from Code Fellows bootcamp with expertise in agile development methodology and experience with software development using continuous integration/deployment pipeline. A motivated team player who is highly proficient in Java and well versed in JavaScript with experience building Android apps.
         </p>
