+++ src/main/resources/templates/contactus.html
@@ -1,0 +1,46 @@
+<!DOCTYPE html>
+<html lang="en">
+<!--Reference: https://bootsnipp.com/snippets/7nmOW-->
+<head>
+    <meta charset="UTF-8">
+    <title>User Profile Page</title>
+    <th:block th:replace="fragments/common :: head"> </th:block>
+</head>
+<body>
+<nav th:replace="fragments/common :: navbar"></nav>
+
+<div class="container contact-form">
+    <div class="contact-image">
+<!--        <img src="https://image.ibb.co/kUagtU/rocket_contact.png" alt="rocket_contact"/>-->
+    </div>
+    <br>
+    <br>
+    <br>
+    <form method="post">
+        <h3>Drop Us a Message</h3>
+        <div class="row">
+            <div class="col-md-6">
+                <div class="form-group">
+                    <input type="text" name="txtName" class="form-control" placeholder="Your Name *" value="" />
+                </div>
+                <div class="form-group">
+                    <input type="text" name="txtEmail" class="form-control" placeholder="Your Email *" value="" />
+                </div>
+                <div class="form-group">
+                    <input type="text" name="txtPhone" class="form-control" placeholder="Your Phone Number *" value="" />
+                </div>
+                <div class="form-group">
+                    <input type="submit" name="btnSubmit" class="btnContact" value="Send Message" />
+                </div>
+            </div>
+            <div class="col-md-6">
+                <div class="form-group">
+                    <textarea name="txtMsg" class="form-control" placeholder="Your Message *" style="width: 100%; height: 150px;"></textarea>
+                </div>
+            </div>
+        </div>
+    </form>
+</div>
+
+</body>
+</html>
