+++ src/main/resources/templates/fragments/common.html
@@ -30,16 +30,16 @@
                 <li class="narbar-item"><a href="/profile" class="nav-link">Profile</a></li>
                 <li class="narbar-item"><a href="/products" class="nav-link">Products</a></li>
                 <li class="narbar-item"><a href="/mycart" class="nav-link">Cart</a></li>
-                <li class="narbar-item"><a href="#" class="nav-link">Contact Us</a></li>
-                <li class="narbar-item"><a href="#" class="nav-link">About Us</a></li>
+                <li class="narbar-item"><a href="/contactus" class="nav-link">Contact Us</a></li>
+                <li class="narbar-item"><a href="/about-us" class="nav-link">About Us</a></li>
                 <li class="narbar-item"><a href="/logout" class="nav-link">Log out</a></li>
 
             </ul>
             <ul th:unless="${username}" class="navbar-nav ml-auto">
                 <li class="narbar-item"><a href="/login" class="nav-link">Sign In</a></li>
                 <li class="narbar-item"><a href="/signup" class="nav-link">Sign up</a></li>
-                <li class="narbar-item"><a href="#" class="nav-link">Contact Us</a></li>
-                <li class="narbar-item"><a href="#" class="nav-link">About Us</a></li>
+                <li class="narbar-item"><a href="/contactus" class="nav-link">Contact Us</a></li>
+                <li class="narbar-item"><a href="/about-us" class="nav-link">About Us</a></li>
             </ul>
         </div>
     </nav>
