+++ src/main/resources/templates/home.html
@@ -9,5 +9,8 @@
 <nav th:replace="fragments/common :: navbar"></nav>
 
 
+
+
+
 </body>
 </html>
