+++ src/main/resources/templates/profile.html
@@ -9,5 +9,11 @@
 <nav th:replace="fragments/common :: navbar"></nav>
 
 
+
+
+
+
+
+
 </body>
 </html>
